refactor: renamed the endpoints to controllers
In order to fit the Spring Boot naming conventions, the enpoints classes
were renamed to controllers
