+++ src/main/java/com/safetynet/alerts/controller/MedicalRecordController.java
@@ -1,4 +1,4 @@
-package com.safetynet.alerts.endpoint;
+package com.safetynet.alerts.controller;
 
 import java.util.Date;
 import java.util.List;
@@ -25,8 +25,8 @@ import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping(value = "/medicalRecord", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-public class MedicalRecordEndpoint {
-    private static final Logger logger = LoggerFactory.getLogger(MedicalRecordEndpoint.class);
+public class MedicalRecordController {
+    private static final Logger logger = LoggerFactory.getLogger(MedicalRecordController.class);
 
     @Autowired
     private MedicalRecordService medicalRecordService;
