+++ src/main/java/com/safetynet/alerts/service/FireStationService.java
@@ -13,7 +13,7 @@ import com.safetynet.alerts.dto.node.MedicalRecordsDTO;
 import com.safetynet.alerts.dto.node.PersonsDTO;
 import com.safetynet.alerts.dto.resource.FireStationDTO;
 import com.safetynet.alerts.dto.response.CoveredPersonResponse;
-import com.safetynet.alerts.repo.DataRepository;
+import com.safetynet.alerts.repository.DataRepository;
 
 @Service
 public class FireStationService {
