+++ src/main/java/com/safetynet/alerts/service/MedicalRecordService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 
 import com.safetynet.alerts.dto.node.MedicalRecordsDTO;
-import com.safetynet.alerts.repo.DataRepository;
+import com.safetynet.alerts.repository.DataRepository;
 
 @Service
 public class MedicalRecordService {
