+++ src/main/java/com/safetynet/alerts/service/PersonInfoService.java
@@ -12,7 +12,7 @@ import com.safetynet.alerts.dto.node.PersonsDTO;
 import com.safetynet.alerts.dto.resource.PersonInfoDTO;
 import com.safetynet.alerts.dto.response.MedicalHistory;
 import com.safetynet.alerts.dto.response.PersonResponse;
-import com.safetynet.alerts.repo.DataRepository;
+import com.safetynet.alerts.repository.DataRepository;
 
 @Service
 public class PersonInfoService {
