+++ src/main/java/com/safetynet/alerts/service/PersonService.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 
 import com.safetynet.alerts.dto.node.PersonsDTO;
-import com.safetynet.alerts.repo.DataRepository;
+import com.safetynet.alerts.repository.DataRepository;
 
 @Service
 public class PersonService {
