+++ src/main/java/com/safetynet/alerts/service/PhoneAlertService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 import com.safetynet.alerts.dto.node.FireStationsDTO;
 import com.safetynet.alerts.dto.node.PersonsDTO;
 import com.safetynet.alerts.dto.resource.PhoneAlertDTO;
-import com.safetynet.alerts.repo.DataRepository;
+import com.safetynet.alerts.repository.DataRepository;
 
 @Service
 public class PhoneAlertService {
