+++ src/test/java/com/safetynet/alerts/e2e/PersonControllerE2ETest.java
@@ -22,20 +22,20 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.safetynet.alerts.endpoint.MedicalRecordEndpoint;
-import com.safetynet.alerts.service.MedicalRecordService;
+import com.safetynet.alerts.controller.PersonController;
+import com.safetynet.alerts.service.PersonService;
 import com.safetynet.alerts.util.NodeConstructorTestUtil;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @TestInstance(Lifecycle.PER_CLASS)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class MedicalRecordEndpointE2ETest {
+public class PersonControllerE2ETest {
 
     @Value(value = "${local.server.port}")
     private int port;
 
     @Autowired
-    MedicalRecordService service;
+    PersonService service;
 
     @Autowired
     private TestRestTemplate restTemplate;
@@ -44,7 +44,7 @@ public class MedicalRecordEndpointE2ETest {
     ObjectMapper objectMapper;
 
     @Autowired
-    MedicalRecordEndpoint medicalRecordEndpoint;
+    PersonController personEndpoint;
 
     private NodeConstructorTestUtil nodeConstructor;
 
@@ -59,59 +59,60 @@ public class MedicalRecordEndpointE2ETest {
 
     @Test
     @Order(1)
-    public void givenValidMedicalRecordAsJson_whenAddMedicalRecord_thenResponseStatusIsCreated() throws Exception {
+    public void givenValidPersonAsJson_whenAddPerson_thenResponseStatusIsCreated() throws Exception {
         // given
-        final String payload = nodeConstructor.createValidMedicalRecordAsJson();
+        final String payload = nodeConstructor.createValidPersonAsJson();
         final HttpEntity<String> request = new HttpEntity<String>(payload, httpHeaders);
         // when
-        final ResponseEntity<Void> response = this.restTemplate.postForEntity("http://localhost:" + port + "/api/medicalRecord", request, Void.class);
+        final ResponseEntity<Void> response = this.restTemplate.postForEntity("http://localhost:" + port + "/api/person", request, Void.class);
         // then
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
     }
 
     @Test
     @Order(2)
-    public void givenInvalidMedicalRecordAsJson_whenAddMedicalRecord_thenResponseStatusIsBadRequest() throws Exception {
+    public void givenInvalidPersonAsJson_whenAddPerson_thenResponseStatusIsBadRequest() throws Exception {
         // given
-        final String payload = nodeConstructor.createInvalidMedicalRecordAsJson();
+        final String payload = nodeConstructor.createInvalidPersonAsJson();
         final HttpEntity<String> request = new HttpEntity<String>(payload, httpHeaders);
         // when
-        final ResponseEntity<Void> response = this.restTemplate.postForEntity("http://localhost:" + port + "/api/medicalRecord", request, Void.class);
+        final ResponseEntity<Void> response = this.restTemplate.postForEntity("http://localhost:" + port + "/api/person", request, Void.class);
         // then
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.BAD_REQUEST);
     }
 
     @Test
     @Order(3)
-    public void givenValidParameters_whenModifyMedicalRecord_thenResponseStatusIsNoContent() throws Exception {
+    public void givenValidParameters_whenModifyPerson_thenResponseStatusIsNoContent() throws Exception {
         // given
-        final HttpEntity<Void> request = new HttpEntity<Void>(httpHeaders);
+        final HttpEntity<String> request = new HttpEntity<String>(httpHeaders);
         // when
         final ResponseEntity<Void> response = this.restTemplate
-                .exchange("http://localhost:" + port + "/api/medicalRecord?firstName=John&lastName=Boyd&allergies=pollen", HttpMethod.PUT, request, Void.class);
+                .exchange("http://localhost:" + port + "/api/person?firstName=John&lastName=Boyd&address=new address", HttpMethod.PUT, request, Void.class);
         // then
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.NO_CONTENT);
     }
 
     @Test
     @Order(4)
-    public void givenInvalidParameters_whenModifyMedicalRecord_thenResponseStatusIsNotFound() throws Exception {
+    public void givenInvalidParameters_whenModifyPerson_thenResponseStatusIsNotFound() throws Exception {
         // given
-        final HttpEntity<Void> request = new HttpEntity<Void>(httpHeaders);
+        final String payload = nodeConstructor.createInvalidPersonAsJson();
+        final HttpEntity<String> request = new HttpEntity<String>(payload, httpHeaders);
         // when
-        final ResponseEntity<Void> response = this.restTemplate.exchange(
-                "http://localhost:" + port + "/api/medicalRecordfirstName=Robin&lastName=Hugues&allergies=pollen", HttpMethod.PUT, request, Void.class);
+        final ResponseEntity<Void> response = this.restTemplate
+                .exchange("http://localhost:" + port + "/api/person?firstName=Invalid&lastName=Name&address=new address", HttpMethod.PUT, request, Void.class);
         // then
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
     }
 
     @Test
     @Order(5)
-    public void givenValidParameters_whenDeleteMedicalRecord_thenResponseStatusIsOk() throws Exception {
+    public void givenValidParameters_whenDeletePerson_thenResponseStatusIsOk() throws Exception {
         // given
-        final HttpEntity<Void> request = new HttpEntity<Void>(httpHeaders);
+        final HttpEntity<String> request = new HttpEntity<String>(httpHeaders);
         // when
-        final ResponseEntity<Void> response = this.restTemplate.exchange("http://localhost:" + port + "/api/medicalRecord?firstName=John&lastName=Boyd",
+        final ResponseEntity<Void> response = this.restTemplate.exchange("http://localhost:" + port + "/api/person?firstName=Little&lastName=Child",
                 HttpMethod.DELETE, request, Void.class);
         // then
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
@@ -119,11 +120,12 @@ public class MedicalRecordEndpointE2ETest {
 
     @Test
     @Order(6)
-    public void givenInvalidParameters_whenDeleteMedicalRecord_thenResponseStatusIsNotFound() throws Exception {
+    public void givenInvalidParameters_whenDeletePerson_thenResponseStatusIsNotFound() throws Exception {
         // given
-        final HttpEntity<Void> request = new HttpEntity<Void>(httpHeaders);
+        final String payload = nodeConstructor.createInvalidPersonAsJson();
+        final HttpEntity<String> request = new HttpEntity<String>(payload, httpHeaders);
         // when
-        final ResponseEntity<Void> response = this.restTemplate.exchange("http://localhost:" + port + "/api/medicalRecordfirstName=Robin&lastName=Hugues",
+        final ResponseEntity<Void> response = this.restTemplate.exchange("http://localhost:" + port + "/api/person?firstName=Invalid&lastName=Name",
                 HttpMethod.DELETE, request, Void.class);
         // then
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
