+++ src/test/java/com/safetynet/alerts/service/ChildAlertServiceUnitTest.java
@@ -15,7 +15,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import com.safetynet.alerts.dto.resource.ChildAlertDTO;
 import com.safetynet.alerts.dto.response.ChildResponse;
 import com.safetynet.alerts.dto.response.FamilyResponse;
-import com.safetynet.alerts.repo.DataRepository;
+import com.safetynet.alerts.repository.DataRepository;
 import com.safetynet.alerts.util.JsonDataRepositoryTestUtil;
 import com.safetynet.alerts.util.TestRepository;
