+++ src/test/java/com/safetynet/alerts/service/FloodServiceUnitTest.java
@@ -16,7 +16,7 @@ import com.safetynet.alerts.dto.resource.FloodDTO;
 import com.safetynet.alerts.dto.response.CoveredHouseResponse;
 import com.safetynet.alerts.dto.response.InhabitantResponse;
 import com.safetynet.alerts.dto.response.MedicalHistory;
-import com.safetynet.alerts.repo.DataRepository;
+import com.safetynet.alerts.repository.DataRepository;
 import com.safetynet.alerts.util.JsonDataRepositoryTestUtil;
 import com.safetynet.alerts.util.TestRepository;
